[PATCH] kNFSd: NFSv4: fix type checking in fh_verify()
Change the type checking in fh_verify().  This fixes a bug which
I reported on the mailing list a few days ago, and also adds a
new error code nfserr_symlink (v4 only).  This is returned whenever
an operation which is illegal for symlinks is attempted on a symlink,
and takes precedence over ERR_NOTDIR or ERR_INVAL.

BKrev: 3d6680c4bljkDs8R7YKGajjwGSBS7Q