[PATCH] kNFSd: NFSv4: allow resfh==fhp in fh_compose()
Change fh_compose() so that it will do the right thing if fhp==res_fh.
(This is convenient in the NFSv4 LOOKUP operation, which _replaces_
CURRENT_FH with the filehandle obtained by lookup.)

This requires extracting the interesting parts of the reference
filehandle first, then calling fh_put if it is a re-use.

BKrev: 3d6680f1w7KLhlPQWSgr03CNBLE42w