[PATCH] have lockd and rpciod drop locks on exit
Neither lockd nor rpciod properly drop the BKL on exit.

Yes, the BKL is automatically relinquished on schedule but the failure
to drop the lock throws off the atomic accounting and results in a
preempt_count warning on exit.

This makes them play fair - it does not hurt.

BKrev: 3d6e82e5JevZKFt616rusP2VY1fcvQ