[PATCH] ptrace-fix-2.5.33-A1
This contains Daniel's suggested fix that allows a parent to
PTRACE_ATTACH to a child it forked.  That fixes the incorrect BUG_ON()
assert that Ogawa's patch was intended to fix, and we thus undo Ogawa's
patch.

I've tested various ptrace uses and they appear to work just fine.

BKrev: 3d777b5dglGrHtGhsUouGOGxqwWVZA