[PATCH] resurrect CONFIG_HIGHPTE
Bill Irwin's patch to fix up pte's in highmem.

With CONFIG_HIGHPTE, the direct pte pointer in struct page becomes the
64-bit physical address of the single pte which is mapping this page.

If the page is not PageDirect then page->pte.chain points at a list of
pte_chains, which each now contain an array of 64-bit physical
addresses of the pte's which are mapping the page.

The functions rmap_ptep_map() and rmap_ptep_unmap() are used for
mapping and unmapping the page which backs the target pte.

The patch touches all architectures (adding do-nothing compatibility
macros and inlines).  It generally mangles lots of header files and may
break non-ia32 compiles.  I've had it in testing since 2.5.31.

BKrev: 3d7e35670yGiHdiUZbo_lAhViOP-vw