Merge bk://linuxusb.bkbits.net/linus-2.5 into home.transmeta.com:/home/torvalds/v2.5/linux
2002/09/12 10:27:47-07:00 jdike
[PATCH] UML arch (user-mode Linux)

This patch implements UML for 2.5.34.

2002/09/12 08:35:49-07:00 adam
[PATCH] Building list of drives in right order

	ata_attach in linux-2.5.34/drivers/ide/ide.c builds a list of
IDE drives that do not yet have a device driver bound to them, in case
ide-disk, ide-scsi, or whatever driver you want to use is not loaded
yet.

	The problem was that ata_attach was adding to the head of
the list, so the list was being built in reverse order.  So, if
you had two IDE disks, and ide-disk was a loadable module, the
devfs entries for the disks would be numbered in reverse (the
first disk would be /dev/discs/disc1, and the second would be
/dev/discs/disc0).

	This fixes the problem by changing the relevant list_add to
list_add_tail.  Incidentally, the generic code in drivers/base/ already
does it this way.

2002/09/11 16:47:16-07:00 torvalds
Allocate system call numbers: 250 and 251 for hugetlb, with
252 for exit_group

BKrev: 3d80d20fObpZY43banHAnE6pWmtJJg