Merge kroah.com:/home/greg/linux/BK/bleeding_edge-2.5 into kroah.com:/home/greg/linux/BK/gregkh-2.5
2002/09/16 10:31:51-07:00 mdharm-usb
[PATCH] USB storage: add error checks, remove useless code

This patch removes attempts to clear halts on a control endpoint (think
about it for a minute if you don't see why this is pointless....) and also
adds return-code checks for all places where halts are cleared.

This _should_ be just redundant code, but recent tests suggest that this
is, in fact, not the case.  People should _heavily_ test this patch.  I'm
going to pause here for a while (in the patch stream) until we've got this
sorted out -- initial results on my test setup seem to show some problems
still remain.  Where those problems are (HCD or usb-storage) remains to be
seen.

2002/09/16 10:31:22-07:00 mdharm-usb
[PATCH] USB storage: minor compilation fixes

This patch fixes up some minor compilation problems.

2002/09/16 10:30:55-07:00 mdharm-usb
[PATCH] USB storage: macro-ize address manipulation

This patch converts all uses of page_addres() to the sg_address() macro.
This will make backporting to 2.4 easier, as well as eliminate lots of
redundant code.

2002/09/16 10:30:27-07:00 mdharm-usb
[PATCH] USB storage: remove tests against EINPROGRESS

This patch removes tests of urb->status for EINPROGRESS.  As was pointed
out, that's not such a good idea, for a variety of reasons.

In the process, a semaphore became useless.

2002/09/16 10:30:06-07:00 stern
[PATCH] USB storage:  Merging raw_bulk.c with transport.c

Here's a very simple patch that can go into the source tree right away.
It just fixes some occurrences of the scsi result code GOOD to GOOD << 1
in isd200.c.

BKrev: 3d8675c0Iq0IOq1IjOnATNgczC4bxw